+++ src/Service/CrowdfundingService.php
@@ -36,29 +36,4 @@ class CrowdfundingService implements CrowdfundingServiceInterface
     {
         return $this->crowdfundingRepository->findAllSorted();
     }
-
-    // Finds all the medias
-    public function findAllMedias()
-    {
-        return $this->crowdfundingMediaRepository->findAll();
-    }
-
-    // Finds one by id
-    public function findOneById(int $id): Crowdfunding
-    {
-        return $this->crowdfundingRepository->findOneById($id);
-    }
-
-    // Searches for crowdfunding
-    public function search(string $query)
-    {
-        return $this->crowdfundingRepository->search($query);
-    }
-
-    // Saves the crowdfunding
-    public function save(Crowdfunding $crowdfunding): void
-    {
-        $this->entityManager->persist($crowdfunding);
-        $this->entityManager->flush();
-    }
 }